% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdalnara-package.R
\docType{package}
\name{gdalnara-package}
\alias{gdalnara}
\alias{gdalnara-package}
\title{gdalnara: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mdsumner/gdalnara}
  \item Report bugs at \url{https://github.com/mdsumner/gdalnara/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
