% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdalnara-package.R
\docType{package}
\name{gdalnara-package}
\alias{gdalnara}
\alias{gdalnara-package}
\title{gdalnara: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
